+++ examples/energy.py
@@ -1,25 +1,25 @@
+from pathlib import Path
+
+from qcio import Molecule, ProgramInput, SinglePointOutput
+
 from chemcloud import CCClient
-from chemcloud.models import AtomicInput, Molecule
+
+current_dir = Path(__file__).resolve().parent
+water = Molecule.open(current_dir / "h2o.xyz")
 
 client = CCClient()
 
-water = Molecule.from_data("pubchem:water")
-atomic_input = AtomicInput(
+prog_inp = ProgramInput(
     molecule=water,
-    model={"method": "B3LYP", "basis": "6-31g"},
-    driver="energy",
-    keywords={
-        "closed": True,
-        "restricted": True,
-    },
-    protocols={"stdout": True, "native_files": "all"},
-    extras={"tcfe:keywords": {"native_files": ["c0"]}},
+    model={"method": "b3lyp", "basis": "6-31gg"},
+    calctype="energy",
+    keywords={},
 )
-future_result = client.compute(atomic_input, engine="terachem_fe")
-result = future_result.get()
-# AtomicResult object containing all returned data
-print(result)
+future_result = client.compute("psi4", prog_inp, collect_files=True)
+output: SinglePointOutput = future_result.get()
+# SinglePointOutput object containing all returned data
+print(output.stdout)
+print(output)
 # The energy value requested
-print(result.return_result)
-print(result.stdout)
-print(result.native_files.keys())
+print(output.return_result)
+print(output.files.keys())