+++ scripts/format.sh
@@ -2,6 +2,6 @@
 
 set -x
 
-autoflake --remove-all-unused-imports --recursive --remove-unused-variables --in-place examples chemcloud tests examples --exclude=__init__.py
-black examples chemcloud tests examples
-isort examples chemcloud tests examples
+poetry run black .
+poetry run isort .
+poetry run ruff check --fix .